refactor: qc-application feed 도메인 패키지 정리
