+++ qc-application/src/main/java/com/eager/questioncloud/application/api/feed/library/controller/FeedLibraryController.java
@@ -1,6 +1,7 @@
-package com.eager.questioncloud.application.api.feed;
+package com.eager.questioncloud.application.api.feed.library.controller;
 
 import com.eager.questioncloud.application.api.common.PagingResponse;
+import com.eager.questioncloud.application.api.feed.library.service.FeedLibraryService;
 import com.eager.questioncloud.core.domain.question.common.QuestionFilter;
 import com.eager.questioncloud.core.domain.userquestion.dto.UserQuestionDetail;
 import io.swagger.v3.oas.annotations.Operation;
